+++ packages/core/src/lib/message/version_0.spec.ts
@@ -1,4 +1,4 @@
-import { proto_message } from "@waku/proto";
+import { WakuMessage } from "@waku/proto";
 import { expect } from "chai";
 import fc from "fast-check";
 
@@ -69,7 +69,7 @@ describe("Waku Message version 0", function () {
           // Encode the length of the payload
           // Not a relevant real life example
           const metaSetter = (
-            msg: proto_message.WakuMessage & { meta: undefined }
+            msg: WakuMessage & { meta: undefined }
           ): Uint8Array => {
             const buffer = new ArrayBuffer(4);
             const view = new DataView(buffer);
@@ -91,7 +91,7 @@ describe("Waku Message version 0", function () {
           )) as DecodedMessage;
 
           const expectedMeta = metaSetter(
-            new proto_message.WakuMessage({
+            new WakuMessage({
               payload,
               timestamp: undefined,
               contentTopic: "",
@@ -99,7 +99,7 @@ describe("Waku Message version 0", function () {
               meta: undefined,
               rateLimitProof: undefined,
               version: undefined,
-            }) as proto_message.WakuMessage & { meta: undefined }
+            }) as WakuMessage & { meta: undefined }
           );
 
           expect(result.meta).to.deep.eq(expectedMeta);
