+++ packages/core/src/lib/store/index.ts
@@ -15,12 +15,11 @@ import all from "it-all";
 import * as lp from "it-length-prefixed";
 import { pipe } from "it-pipe";
 import { Uint8ArrayList } from "uint8arraylist";
+import { v4 as uuid } from "uuid";
 
 import { BaseProtocol } from "../base_protocol.js";
 import { DefaultPubSubTopic } from "../constants.js";
 import { toProtoMessage } from "../to_proto_message.js";
-
-import { createQuery, PageDirection, Params } from "./history_rpc.js";
 
 const log = debug("waku:store");
 
@@ -28,11 +27,77 @@ export const StoreCodec = "/vac/waku/store/2.0.0-beta4";
 
 export const DefaultPageSize = 10;
 
-export { PageDirection };
-
 export interface TimeFilter {
   startTime: Date;
   endTime: Date;
+}
+
+const OneMillion = BigInt(1_000_000);
+
+export enum PageDirection {
+  BACKWARD = "backward",
+  FORWARD = "forward",
+}
+
+export interface Params {
+  contentTopics: string[];
+  pubSubTopic: string;
+  pageDirection: PageDirection;
+  pageSize: number;
+  startTime?: Date;
+  endTime?: Date;
+  cursor?: proto.Index;
+}
+
+export function createQuery(params: Params): proto.HistoryRpc {
+  const contentFilters = params.contentTopics.map((contentTopic) => {
+    return { contentTopic };
+  });
+
+  const direction = directionToProto(params.pageDirection);
+
+  const pagingInfo = {
+    pageSize: BigInt(params.pageSize),
+    cursor: params.cursor,
+    direction,
+  } as proto.PagingInfo;
+
+  let startTime, endTime;
+  if (params.startTime) {
+    // milliseconds 10^-3 to nanoseconds 10^-9
+    startTime = BigInt(params.startTime.valueOf()) * OneMillion;
+  }
+
+  if (params.endTime) {
+    // milliseconds 10^-3 to nanoseconds 10^-9
+    endTime = BigInt(params.endTime.valueOf()) * OneMillion;
+  }
+
+  const query = new proto.HistoryQuery({
+    pubsubTopic: params.pubSubTopic,
+    contentFilters,
+    pagingInfo,
+    startTime,
+    endTime,
+  });
+
+  return new proto.HistoryRpc({
+    requestId: uuid(),
+    query,
+  });
+}
+
+function directionToProto(
+  pageDirection: PageDirection
+): proto.PagingInfo_Direction {
+  switch (pageDirection) {
+    case PageDirection.BACKWARD:
+      return proto.PagingInfo_Direction.BACKWARD;
+    case PageDirection.FORWARD:
+      return proto.PagingInfo_Direction.FORWARD;
+    default:
+      return proto.PagingInfo_Direction.BACKWARD;
+  }
 }
 
 export interface QueryOptions {
