+++ packages/core/src/lib/to_proto_message.spec.ts
@@ -1,4 +1,4 @@
-import { proto_message } from "@waku/proto";
+import { WakuMessage } from "@waku/proto";
 import { expect } from "chai";
 
 import { toProtoMessage } from "./to_proto_message.js";
@@ -6,7 +6,7 @@ import { toProtoMessage } from "./to_proto_message.js";
 //this test will fail as the initalisation to wire will drop the fields that are undefined
 describe.skip("to proto message", () => {
   it("Fields are not dropped", () => {
-    const wire: proto_message.WakuMessage = new proto_message.WakuMessage({
+    const wire: WakuMessage = new WakuMessage({
       payload: new Uint8Array(),
       contentTopic: "foo",
     });
