+++ packages/proto/src/index.ts
@@ -1,6 +1,14 @@
 export * as proto_filter from "./gen/filter_pb.js";
+
 export * as proto_message from "./gen/message_pb.js";
+export { WakuMessage } from "./gen/message_pb.js";
+
 export * as proto_lightpush from "./gen/light_push_pb.js";
+export { PushResponse } from "./gen/light_push_pb.js";
+
 export * as proto_peer_exchange from "./gen/peer_exchange_pb.js";
+
 export * as proto_store from "./gen/store_pb.js";
+
 export * as proto_topic_only_message from "./gen/topic_only_message_pb.js";
+export { TopicOnlyMessage } from "./gen/topic_only_message_pb.js";
